requirements.txt
github3.py==4.0.1
python-dotenv==1.0.1
requests==2.31.0
python-dotenv==1.0.1